requirements.txt
asgiref==3.8.1
boto3==1.37.37
botocore==1.37.37
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.2
django-cors-headers==4.7.0
django-getenv==1.3.2
django-storages==1.14.6
djangorestframework==3.16.0
drf-nested-routers==0.94.1
ecdsa==0.19.1
idna==3.10
jmespath==1.0.1
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
requests==2.32.3
rsa==4.9.1
s3transfer==0.11.5
six==1.17.0
sqlparse==0.5.3
urllib3==2.4.0
